rename package to autoware_overlay_rviz_plugin
