+++ common/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/assets/images/select_add.png
