+++ common/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_vehicle_overlay_rviz_plugin</name>
+  <name>autoware_overlay_rviz_plugin</name>
   <version>0.0.1</version>
   <description>
     RViz2 plugin for 2D overlays in the 3D view. Mainly a port of the JSK overlay plugin
