+++ common/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/src/overlay_text_display.cpp
@@ -67,7 +67,7 @@
 #include <regex>
 #include <sstream>
 
-namespace autoware_vehicle_overlay_rviz_plugin
+namespace autoware_overlay_rviz_plugin
 {
 OverlayTextDisplay::OverlayTextDisplay()
 : texture_width_(0),
@@ -212,7 +212,7 @@ void OverlayTextDisplay::update(float /*wall_dt*/, float /*ros_dt*/)
 
   overlay_->updateTextureSize(texture_width_, texture_height_);
   {
-    autoware_vehicle_overlay_rviz_plugin::ScopedPixelBuffer buffer = overlay_->getBuffer();
+    autoware_overlay_rviz_plugin::ScopedPixelBuffer buffer = overlay_->getBuffer();
     QImage Hud = buffer.getQImage(*overlay_, bg_color_);
     QPainter painter(&Hud);
     painter.setRenderHint(QPainter::Antialiasing, true);
@@ -297,7 +297,7 @@ void OverlayTextDisplay::processMessage(autoware_overlay_msgs::msg::OverlayText:
     static int count = 0;
     std::stringstream ss;
     ss << "OverlayTextDisplayObject" << count++;
-    overlay_.reset(new autoware_vehicle_overlay_rviz_plugin::OverlayObject(ss.str()));
+    overlay_.reset(new autoware_overlay_rviz_plugin::OverlayObject(ss.str()));
     overlay_->show();
   }
   if (overlay_) {
@@ -550,8 +550,8 @@ void OverlayTextDisplay::updateLineWidth()
   }
 }
 
-}  // namespace autoware_vehicle_overlay_rviz_plugin
+}  // namespace autoware_overlay_rviz_plugin
 
 #include <pluginlib/class_list_macros.hpp>
 PLUGINLIB_EXPORT_CLASS(
-  autoware_vehicle_overlay_rviz_plugin::OverlayTextDisplay, rviz_common::Display)
+  autoware_overlay_rviz_plugin::OverlayTextDisplay, rviz_common::Display)
