+++ common/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/src/steering_wheel_display.cpp
@@ -31,14 +31,14 @@
 #include <memory>
 #include <string>
 
-namespace autoware_vehicle_overlay_rviz_plugin
+namespace autoware_overlay_rviz_plugin
 {
 
 SteeringWheelDisplay::SteeringWheelDisplay()
 {
   // Load the Quicksand font
   std::string package_path =
-    ament_index_cpp::get_package_share_directory("autoware_vehicle_overlay_rviz_plugin");
+    ament_index_cpp::get_package_share_directory("autoware_overlay_rviz_plugin");
   std::string font_path = package_path + "/assets/font/Quicksand/static/Quicksand-Regular.ttf";
   std::string font_path2 = package_path + "/assets/font/Quicksand/static/Quicksand-Bold.ttf";
   int fontId = QFontDatabase::addApplicationFont(
@@ -121,4 +121,4 @@ QImage SteeringWheelDisplay::coloredImage(const QImage & source, const QColor & 
   return result;
 }
 
-}  // namespace autoware_vehicle_overlay_rviz_plugin
+}  // namespace autoware_overlay_rviz_plugin
